+++ packages/ast-spec/src/element/AccessorProperty/fixtures/key-number/fixture.ts
@@ -1,0 +1,3 @@
+class Foo {
+  accessor 1 = 2;
+}