+++ packages/ast-spec/src/parameter/RestElement/spec.ts
@@ -8,8 +8,8 @@ import type { AssignmentPattern } from '../AssignmentPattern/spec';
 export interface RestElement extends BaseNode {
   type: AST_NODE_TYPES.RestElement;
   argument: DestructuringPattern;
-  typeAnnotation?: TSTypeAnnotation;
-  optional?: boolean;
-  value?: AssignmentPattern;
-  decorators?: Decorator[];
+  typeAnnotation: TSTypeAnnotation | undefined;
+  optional: boolean;
+  value: AssignmentPattern | undefined;
+  decorators: Decorator[];
 }