+++ packages/eslint-plugin/src/rules/no-invalid-this.ts
@@ -20,13 +20,9 @@ export default createRule<Options, MessageIds>({
     docs: {
       description:
         'Disallow `this` keywords outside of classes or class-like objects',
-      recommended: false,
       extendsBaseRule: true,
     },
-    // TODO: this rule has only had messages since v7.0 - remove this when we remove support for v6
-    messages: baseRule.meta.messages ?? {
-      unexpectedThis: "Unexpected 'this'.",
-    },
+    messages: baseRule.meta.messages,
     hasSuggestions: baseRule.meta.hasSuggestions,
     schema: baseRule.meta.schema,
   },