+++ packages/parser/tests/fixtures/scope-analysis/class-supper-type.ts
@@ -1,11 +1,5 @@
-abstract class Foo extends Bar<Baz> {
+abstract class Foo extends Bar<Baz> {}
 
-}
+declare class Foo2 extends Bar<Baz> {}
 
-declare class Foo2 extends Bar<Baz> {
-
-}
-
-class Foo3 extends Bar<Baz> {
-
-}
+class Foo3 extends Bar<Baz> {}