+++ packages/parser/tests/fixtures/scope-analysis/expression-type-parameters.ts
@@ -1,4 +1,4 @@
-const a, b, c, d, e, f
+const a, b, c, d, e, f;
 
 new foo<Bar>(a, b, c);