+++ packages/parser/tests/lib/parser.ts
@@ -1,6 +1,6 @@
-import * as scopeManager from '@typescript-eslint/scope-manager/dist/analyze';
+import * as scopeManager from '@typescript-eslint/scope-manager';
 import type { ParserOptions } from '@typescript-eslint/types';
-import * as typescriptESTree from '@typescript-eslint/typescript-estree/dist/parser';
+import * as typescriptESTree from '@typescript-eslint/typescript-estree';
 
 import { parse, parseForESLint } from '../../src/parser';
 
@@ -19,11 +19,6 @@ describe('parser', () => {
     expect(() => parseForESLint(code, null)).not.toThrow();
   });
 
-  it("parseForESLint() should work if options.ecmaVersion is `'latest'`", () => {
-    const code = 'const valid = true;';
-    expect(() => parseForESLint(code, { ecmaVersion: 'latest' })).not.toThrow();
-  });
-
   it('parseAndGenerateServices() should be called with options', () => {
     const code = 'const valid = true;';
     const spy = jest.spyOn(typescriptESTree, 'parseAndGenerateServices');
@@ -33,7 +28,6 @@ describe('parser', () => {
       range: false,
       tokens: false,
       sourceType: 'module' as const,
-      ecmaVersion: 2018,
       ecmaFeatures: {
         globalReturn: false,
         jsx: false,
@@ -41,7 +35,6 @@ describe('parser', () => {
       // ts-estree specific
       filePath: 'isolated-file.src.ts',
       project: 'tsconfig.json',
-      errorOnUnknownASTType: false,
       errorOnTypeScriptSyntacticAndSemanticIssues: false,
       tsconfigRootDir: 'tests/fixtures/services',
       extraFileExtensions: ['.foo'],
@@ -84,7 +77,6 @@ describe('parser', () => {
       range: false,
       tokens: false,
       sourceType: 'module' as const,
-      ecmaVersion: 2018,
       ecmaFeatures: {
         globalReturn: false,
         jsx: false,
@@ -96,7 +88,6 @@ describe('parser', () => {
       // ts-estree specific
       filePath: 'isolated-file.src.ts',
       project: 'tsconfig.json',
-      errorOnUnknownASTType: false,
       errorOnTypeScriptSyntacticAndSemanticIssues: false,
       tsconfigRootDir: 'tests/fixtures/services',
       extraFileExtensions: ['.foo'],
@@ -104,7 +95,6 @@ describe('parser', () => {
     parseForESLint(code, config);
     expect(spy).toHaveBeenCalledTimes(1);
     expect(spy).toHaveBeenLastCalledWith(expect.anything(), {
-      ecmaVersion: 2018,
       globalReturn: false,
       lib: ['dom.iterable'],
       jsxPragma: 'Foo',