+++ packages/typescript-estree/tests/lib/parse.project-true.test.ts
@@ -42,7 +42,7 @@ describe('parseAndGenerateServices', () => {
           filePath: join(PROJECT_DIR, 'notIncluded.ts'),
         }),
       ).toThrow(
-        /project was set to `true` but couldn't find any tsconfig.json relative to '.+\/tests\/fixtures\/projectTrue\/notIncluded.ts' within '.+\/tests\/fixtures\/projectTrue'./,
+        /project was set to `true` but couldn't find any tsconfig.json relative to '.+[/\\]tests[/\\]fixtures[/\\]projectTrue[/\\]notIncluded.ts' within '.+[/\\]tests[/\\]fixtures[/\\]projectTrue'./,
       );
     });
   });